+++ filesys/util/DiretorioUtil.java
@@ -2,9 +2,6 @@ package filesys.util;
 
 import java.util.Map;
 import java.util.Set;
-import java.util.StringTokenizer;
-
-import exception.CaminhoNaoEncontradoException;
 import exception.OperacaoInvalidaException;
 import exception.PermissaoException;
 import filesys.Diretorio;
@@ -13,50 +10,6 @@ import filesys.Usuario;
 public class DiretorioUtil {
 
   private DiretorioUtil() {
-  }
-
-  public static Diretorio obterDiretorioPai(String caminho, Diretorio root, String usuario)
-      throws CaminhoNaoEncontradoException, PermissaoException {
-    String caminhoPai = caminho.substring(0, caminho.lastIndexOf('/'));
-    return navegar(caminhoPai.isEmpty() ? "/" : caminhoPai, root, usuario);
-  }
-
-  
-
-  public static Diretorio navegar(String caminho, Diretorio root, String usuario)
-      throws CaminhoNaoEncontradoException, PermissaoException {
-
-    if (caminho.equals("/") || caminho.isEmpty()) {
-      VerificacaoUtil.verificarPermissaoExecucao(root, usuario, caminho);
-      return root;
-    }
-
-    Diretorio atual = root;
-    StringTokenizer tokenizer = new StringTokenizer(caminho, "/");
-    StringBuilder caminhoAtual = new StringBuilder("/");
-
-    while (tokenizer.hasMoreTokens()) {
-      String parte = tokenizer.nextToken();
-
-      if (!atual.getFilhos().containsKey(parte)) {
-        throw new CaminhoNaoEncontradoException("Caminho não encontrado: " + caminho);
-      }
-
-      atual = atual.getFilhos().get(parte);
-      caminhoAtual.append(parte).append("/");
-      VerificacaoUtil.verificarPermissaoExecucao(atual, usuario, caminhoAtual.toString());
-    }
-
-    return atual;
-  }
-
-  public static Diretorio obterAlvo(Diretorio pai, String nomeAlvo, String caminho)
-      throws CaminhoNaoEncontradoException {
-    Diretorio alvo = pai.getFilhos().get(nomeAlvo);
-    if (alvo == null) {
-      throw new CaminhoNaoEncontradoException("Caminho não encontrado: " + caminho);
-    }
-    return alvo;
   }
 
   public static Diretorio criarDiretorioFilho(Diretorio atual, String nome, String caminhoAtual, String usuario,
@@ -80,7 +33,7 @@ public class DiretorioUtil {
     if (filho.isArquivo()) {
       throw new OperacaoInvalidaException("Não é possível criar um diretório dentro de um arquivo.");
     }
-    VerificacaoUtil.verificarPermissaoExecucao(filho, filho.getDono(), nome); 
+    VerificacaoUtil.verificarPermissaoExecucao(filho, filho.getDono(), nome);
 
     return filho;
   }
@@ -93,11 +46,6 @@ public class DiretorioUtil {
       }
     }
     dir.getFilhos().clear();
-  }
-
-  public static String extrairCaminhoPai(String caminho) {
-    int lastSlash = caminho.lastIndexOf('/');
-    return (lastSlash <= 0) ? "/" : caminho.substring(0, lastSlash);
   }
 
   public static String extrairNomeArquivo(String caminho) {
@@ -134,7 +82,4 @@ public class DiretorioUtil {
     return output.toString();
   }
 
-  public static String obterNomeAlvo(String caminho) {
-    return caminho.substring(caminho.lastIndexOf('/') + 1);
-  }
 }
