+++ filesys/util/VerificacaoUtil.java
@@ -10,26 +10,26 @@ public class VerificacaoUtil {
   private VerificacaoUtil() {
   }
 
+  private static void verificarPermissao(Diretorio dir, String usuario, char tipoPermissao, String mensagemErro)
+      throws PermissaoException {
+    if (!dir.temPermissao(usuario, tipoPermissao)) {
+      throw new PermissaoException(mensagemErro);
+    }
+  }
+
   public static void verificarPermissaoEscrita(Diretorio diretorio, String usuario, String caminhoAtual)
       throws PermissaoException {
-    if (!diretorio.temPermissao(usuario, 'w')) {
-      throw new PermissaoException("Sem permissão para criar em: " + caminhoAtual);
-    }
+    verificarPermissao(diretorio, usuario, 'w', "Sem permissão de escrita para acessar: " + caminhoAtual);
   }
 
   public static void verificarPermissaoExecucao(Diretorio diretorio, String usuario, String caminhoAtual)
       throws PermissaoException {
-    if (!diretorio.temPermissao(usuario, 'x')) {
-      throw new PermissaoException("Sem permissão de execução para acessar: " + caminhoAtual);
-    }
+    verificarPermissao(diretorio, usuario, 'x', "Sem permissão de execução para acessar: " + caminhoAtual);
   }
 
-
-  public static void verificarLeituraDiretorio(Diretorio dir, String usuario, String caminho)
+  public static void verificarPermissaoLeitura(Diretorio dir, String usuario, String caminho)
       throws PermissaoException {
-    if (!dir.temPermissao(usuario, 'r')) {
-      throw new PermissaoException("Sem permissão de leitura para listar o diretório: " + caminho);
-    }
+    verificarPermissao(dir, usuario, 'r', "Sem permissão de leitura para acessar: " + caminho);
   }
 
   public static void verificarRemocaoDiretorio(Diretorio alvo, boolean recursivo) throws PermissaoException {
